(docs) Update issue reference in template
This commit updates the issue reference section
for new feature / enhancement issue template to
instead of mentioning closing a issue to instead show
a list of issues they can provide.

This is done due to there not being any reason that a
new issue to close an existing issue.
Switching to a list of issues is done as it will show up
better if multiple issues are added.