(maint) Update Scriban template for GRM
The original change that was added to the Scriban template added the
`Fix -" prefix to an issue when the label is "Bug", this worked during
testing, but failed when deployed to production :-(

We use a feature within GitReleaseManager to change Bug to Bug Fix, or
Bug Fixes, depending on how many issues there are within this label.

As a result, we need to extend the template change to create an OR
between Bug, Bug Fix, and Bug Fixes.