(GRM) Update to latest template
With the release of GitReleaseManager 0.16.0, the default template has
been updated to include different template parameter names. This
change was required, since 0.16.0 introduced support for GitLab, and
additional template properties were needed to accommodate this.

Once this commit is made, the automated processed on the .github
repository will likely cause a number of PR's to be created, but these
PR's should only be merged once it is known the the target repository
is using Chocolatey.Cake.Recipe 0.26.0 or higher.  It is only in this
version and above that GRM 0.16.0 will be in use.