(#41) Add office use only section to PR template
This commit adds to the PR template to include some "office use only"
sections. These sections are intended for Chocolatey team members to
ensure we cover our bases with our PRs.

This also adds a section to allow explanation for a Draft PR, so we
don't need to go looking through the PR to understand why it's in a
draft state.