(feat) Add new GRM template
This new contributors template will be used to create release notes
which includes contributor information, i.e. who created the issue, and
who submitted the PR, etc.

There is a slight modification here to the default contributor template
that ships with GRM, but that is just to account for how Chocolatey
prefers to have release notes read.