(maint) Update GRM Template
These changes were originally made in this PR:

https://github.com/chocolatey/choco/pull/3754

And the full details of the changes as to why they were made can be
found in the commit message over there.  The changes should have
been made in this repository, so that the syncing could have been
done across mulitple repositories. They are being made in this
repository now, so that things can continue to update correctly.