(maint) Overhaul of labels.yml used for syncing
- Make sure all descriptions end with a ".".
- Switch to using "Triaging", rather than "_Triaging".  The latter has
been added as an alias, so that any existing label will be updated to
new label.
- Feature, Improvement, and Enhancement have been consolidated into a
single label - Enhancement.  Aliases have been added to ensure that
everything is brought over to new label.
- New labels has been added for "Deprecate", "Priority - Urgent", "Area
- Refactor", and "Priority - Normal"
- Fixed typo - "vulnerabilites" -> "vulnerabilities"
- "TaskItem" has been renamed to "WorkItem".  Task _can't_ be used,
since this causes the workflow to error out completely, as Task is a
protected keyword.
- Priority labels have been renamed to more closely match what is used
elsewhere, and aliases have been added to bring over existing instances
of labels.